// src/emission/enum/net-zero-use-of-sold-products.enum.ts
export enum UseOfSoldProductsMethod {
    DIRECT_ENERGY = 'DIRECT_ENERGY',
    DIRECT_COMBUSTED = 'DIRECT_COMBUSTED',
    DIRECT_GREENHOUSE = 'DIRECT_GREENHOUSE',
    INDIRECT_ENERGY = 'INDIRECT_ENERGY',
    INTERMEDIATE_PRODUCTS = 'INTERMEDIATE_PRODUCTS'
}

export enum UseOfSoldProductsTypeNames {
    direct_fuel = 'Direct fuel consumption',
    direct_electricity = 'Direct electricity consumption',
    direct_refrigerant = 'Direct refrigerant leakage',
    direct_combusted = 'Direct combusted',
    direct_greenhouse = 'Direct Greenhouse',
    indirect_fuel = 'Indirect fuel consumption',
    indirect_electricity = 'Indirect electricity consumption',
    indirect_refrigerant = 'Indirect refrigerant leakage',
    indirect_ghg = 'Indirect GHG emitted',
    intermediate = 'Intermediate products'
}